move out dagidx
